% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_circos_plot.R
\name{draw_circos_plot}
\alias{draw_circos_plot}
\title{Drawing circos plot}
\usage{
draw_circos_plot(data, output.file = NULL, width = 2400,
  height = 2400, res = 120, tracks = c("common", "unique_1",
  "unique_2", "divergence_major_allele", "divergence_minor_allele"),
  tracks.names = c(NULL, NULL, NULL, NULL, NULL),
  base.color = c("lightgreen", "lightpink"), point.size = 0.1,
  color.unplaced = FALSE, color.palette = c(`0` = "dodgerblue3", `1` =
  "purple3", `2` = "grey20", `3` = "firebrick2", `4` = "lightblue"),
  sector.titles.expand = NULL)
}
\arguments{
\item{data}{Generated by load_data_files function}

\item{tracks}{Choose tracks to plot from c("common", "unique_1", "unique_2", "divergence_major_allele", "divergence_minor_allele")}

\item{tracks.names}{tracks' name to change}

\item{base.color}{color for background}

\item{point.size}{point size (default: 0.1)}

\item{color.palette}{point color}

\item{bg.col}{background color(default: "white")}

\item{top.track}{to check whether is the top track (defaul: FALSE)}
}
\description{
Drawing circos plot
}
