% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_files.R
\name{load_chromosomes_names}
\alias{load_chromosomes_names}
\title{Loading chromosomes name file}
\usage{
load_chromosomes_names(input_file_path)
}
\arguments{
\item{input_file_path}{Path to chromosomes name file}
}
\value{
Data.frame of chromosomes name file
}
\description{
Two column should provide in this file: original name and corresponding names for sector name
}
