% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_files.R
\name{load_contig_lengths}
\alias{load_contig_lengths}
\title{Loading a single data file generated by postpsass}
\usage{
load_contig_lengths(input_file_path, chromosomes_names = NULL,
  plot.unplaced = TRUE)
}
\arguments{
\item{input_file_path}{file generated by postpsass}

\item{chromosomes_names}{Generated by load_chromosomes_names_file}

\item{plot.unplaced}{Whether to plot unplaced scaffolds}
}
\value{
Output data frame
}
\description{
Loading a single file generated by postpsass
}
