% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_files.R
\name{load_data_files}
\alias{load_data_files}
\title{Loading data file}
\usage{
load_data_files(contig_lengths_file_path, input_file_path, common = TRUE,
  unique_1 = TRUE, unique_2 = TRUE, divergence_minor_allele = TRUE,
  divergence_major_allele = TRUE, chromosomes_names_file_path = NULL,
  plot.unplaced = TRUE)
}
\arguments{
\item{contig_lengths_file_path}{Path to contig lengths file}

\item{input_file_path}{Path to file generated by postpsass}

\item{common}{Whether to load common category data (default: TURE)}

\item{unique_1}{Whether to load unique_2 category data (default: TURE)}

\item{unique_2}{Whether to load unique_2 category data (default: TURE)}

\item{divergence_minor_allele}{Whether to load divergence_minor_allele category data (default: TURE)}

\item{divergence_major_allele}{Whether to load divergence_major_allele category data (default: TURE)}

\item{chromosomes_names_file_path}{Path to chromosomes names file}

\item{plot.unplaced}{Whether to plot unplaced scaffolds}
}
\value{
Output all data in to a list
}
\description{
Loading all data
}
