% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data_files.R
\name{load_single_data_file}
\alias{load_single_data_file}
\title{Loading a single data file generated by postpsass}
\usage{
load_single_data_file(input_file_path, contig_lengths,
  plot.unplaced = TRUE, snp_pos = FALSE)
}
\arguments{
\item{input_file_path}{file generated by postpsass}

\item{contig_lengths}{Generated by load_contig_lengths function}

\item{plot.unplaced}{Whether to plot unplaced scaffolds}
}
\value{
Output data frame
}
\description{
Loading a single file generated by postpsass
}
